Add files via upload
CMOS VLSI Design 4e: A circuits and systems perspective 
The fourth edition of the best-selling text details the modern techniques for the design of complex and high-performance CMOS systems on a chip. Covering the fundamentals of CMOS design from the digital systems level to the circuit level, this book explains the fundamental principles and is a guide to good design practices.